+++ src/mundo/Enemigo.java
@@ -4,10 +4,11 @@ import java.util.Formatter;
 
 public abstract class Enemigo implements SerViviente, Cloneable {
 	/**
-	 * cadena de caracteres incambiable que representa un estado que posee todo enemigo
+	 * cadena de caracteres incambiable que representa un estado que posee todo
+	 * enemigo
 	 */
 	public static final String ATACANDO = "atacando";
-	
+
 	/**
 	 * todo enemigo se rige por posiciones representativas de aparicion
 	 */
@@ -16,7 +17,7 @@ public abstract class Enemigo implements SerViviente, Cloneable {
 	 * todo enemigo se rige por posiciones representativas de ataque
 	 */
 	public static final short POS_ATAQUE = 460;
-	
+
 	/**
 	 * todo enemigo tiene una posicion en el eje Y
 	 */
@@ -37,28 +38,30 @@ public abstract class Enemigo implements SerViviente, Cloneable {
 	 * numero que representa la salud del enemigo
 	 */
 	private byte salud;
-	
+
 	/**
 	 * constructor basico que inicializa unicamente la posicion en Y
 	 */
-	public Enemigo () {
+	public Enemigo() {
 		posY = POS_INICIAL;
 	}
-	
+
 	/**
 	 * Constructor compuesto para cargar informacion
+	 *
 	 * @param posY
 	 * @param estadoActual
 	 * @param frameActual
 	 */
-	public Enemigo (int posY, String estadoActual, byte frameActual) {
-	this.posY = posY;
-	this.estadoActual = estadoActual;
-	this.frameActual = frameActual;
+	public Enemigo(int posY, String estadoActual, byte frameActual) {
+		this.posY = posY;
+		this.estadoActual = estadoActual;
+		this.frameActual = frameActual;
 	}
-	
+
 	/**
 	 * cadena de caracteres que representa el estado actual del zombie
+	 * 
 	 * @return estadoActual
 	 */
 	public String getEstadoActual() {
@@ -67,15 +70,17 @@ public abstract class Enemigo implements SerViviente, Cloneable {
 
 	/**
 	 * cambia el estado actual del enemigo
+	 * 
 	 * @param estadoActual
 	 */
 	public void setEstadoActual(String estadoActual) {
 		this.estadoActual = estadoActual;
-		frameActual= 0;
+		frameActual = 0;
 	}
 
 	/**
 	 * obtiene el valor numerico de la imagen actual con corde al estado
+	 * 
 	 * @return frameActual
 	 */
 	public byte getFrameActual() {
@@ -84,6 +89,7 @@ public abstract class Enemigo implements SerViviente, Cloneable {
 
 	/**
 	 * cambia el frame o numero de la imagen con respecto al estado actual
+	 * 
 	 * @param frameActual
 	 */
 	public void setFrameActual(byte frameActual) {
@@ -92,12 +98,14 @@ public abstract class Enemigo implements SerViviente, Cloneable {
 
 	/**
 	 * obtiene la posicion X del enemigo
+	 * 
 	 * @return posicionX
 	 */
 	public abstract int getPosX();
 
 	/**
 	 * obtiene la posicion Y del enemigo
+	 * 
 	 * @return posY
 	 */
 	public int getPosY() {
@@ -106,6 +114,7 @@ public abstract class Enemigo implements SerViviente, Cloneable {
 
 	/**
 	 * cambia la posicion en el eje Y
+	 *
 	 * @param posY
 	 */
 	public void setPosY(int posY) {
@@ -113,7 +122,9 @@ public abstract class Enemigo implements SerViviente, Cloneable {
 	}
 
 	/**
-	 * obtiene la lentitud del enemigo (tiempo de espera en milisegundos por cada movimiento)
+	 * obtiene la lentitud del enemigo (tiempo de espera en milisegundos por cada
+	 * movimiento)
+	 *
 	 * @return lentitud
 	 */
 	public short getLentitud() {
@@ -122,15 +133,18 @@ public abstract class Enemigo implements SerViviente, Cloneable {
 
 	/**
 	 * cambia la lentitud del enemigo
+	 *
 	 * @param lentitud
 	 */
 	public void setLentitud(short lentitud) {
 		this.lentitud = lentitud;
 	}
+
 	@Override
 	public byte getSalud() {
 		return salud;
 	}
+
 	@Override
 	public void setSalud(byte salud) {
 		this.salud = salud;
@@ -138,38 +152,42 @@ public abstract class Enemigo implements SerViviente, Cloneable {
 
 	/**
 	 * crea una posicion aleatoria en el eje X para la aparicion del enemigo
+	 *
 	 * @return posAleatoria
 	 */
 	public short posAleatoriaX() {
 		int posAleatoria = (int) (Math.random() * SurvivorCamp.ANCHO_PANTALLA/3) + SurvivorCamp.ANCHO_PANTALLA/3 - 75;
 		return (short) posAleatoria;
 	}
-	
+
 	/**
 	 * ejecuta cierta accion al terminar de atacar al personaje
 	 */
 	//public abstract void terminaDeAtacar();
-	
+
 	/**
 	 * comprueba que las posiciones de la bala coincidan con la posicion del enemigo
 	 * y reduzca su salud en caso de ser afectado
+	 *
 	 * @param x
 	 * @param y
 	 * @param danio
 	 * @return true si fue afectado por el disparo
 	 */
-	public abstract boolean comprobarDisparo(int x, int y, byte danio);
-	
+	public abstract boolean comprobarDisparo(int x, int y, int danio);
+
 	/**
 	 * obtiene la ruta de la imagen desde la carpeta img
+	 *
 	 * @return
 	 */
 	public String getURL() {
-		try (Formatter formato = new Formatter()){
-			return "/img/" + getClass().getSimpleName() + "/" + estadoActual + "/" + formato.format("%02d",getFrameActual()) + ".png";
+		try (Formatter formato = new Formatter()) {
+			return "/img/" + getClass().getSimpleName() + "/" + estadoActual + "/"
+					+ formato.format("%02d", getFrameActual()) + ".png";
 		}
 	}
-	
+
 	public Enemigo clonar() {
 		try {
 			return (Enemigo) super.clone();
